+++ src/utils/utils.ts
@@ -1,7 +1,7 @@
-import {TNumber} from "../types/element";
+import {TElement, TNumber} from "../types/element";
 import {Dispatch, SetStateAction} from "react";
 import {ElementStates} from "../types/element-states";
-import {SHORT_DELAY_IN_MS} from "../constants/delays";
+import {DELAY_IN_MS, SHORT_DELAY_IN_MS} from "../constants/delays";
 
 export const setDelay = (delay: number) => {
   return new Promise(resolve => setTimeout(resolve, delay));
@@ -9,6 +9,35 @@ export const setDelay = (delay: number) => {
 
 export const swap = (array: any[], i: number, j: number) => {
   ([array[i], array[j]] = [array[j], array[i]]);
+}
+
+export const unwrapString = async (
+  inputValue: string,
+  setLoader: Dispatch<SetStateAction<boolean>>,
+  setArray: Dispatch<SetStateAction<TElement[]>>,) => {
+
+  setLoader(true);
+  const array: TElement[] = inputValue.split('').map((value: string) => {
+    return {value, color: ElementStates.Default};
+  });
+  setArray(array);
+  const medium = Math.ceil(array.length / 2);
+
+  for (let i = 0; i < medium; i++) {
+    let j = array.length - 1 - i;
+    if (i !== j) {
+      array[i].color = ElementStates.Changing;
+      array[j].color = ElementStates.Changing;
+      setArray([...array]);
+      await setDelay(DELAY_IN_MS);
+    }
+
+    swap(array, i, j);
+    array[i].color = ElementStates.Modified;
+    array[j].color = ElementStates.Modified;
+    setArray([...array]);
+  }
+  setLoader(false);
 }
 
 export const getFibonacciArray = (n: number) => {
